fix: adapt rc.30 merge to 64-bit quota balance extension (WIP)
- merge upstream v1.0.0-rc.30 into main (core wins per sync discipline)
- keep local extensions: 64-bit wallet domain, admin quota adjustment,
  org discount snapshot, invoice reconciliation, durable task journals
- port upstream SettleBilling/LogTaskConsumption onto local types
- restore local task_billing/billing_session/funding_source architectures
- adopt upstream task_polling/relay_task/mjproxy flows where core
- KNOWN WIP: ~46 local-extension tests failing (org discount / invoice /
  MJ billing / task refund) pending architecture ownership decisions;
  Kling e2e double-reserve under mixed flow

+++ common/quota_math_test.go
@@ -131,7 +131,7 @@ func TestQuotaFromDecimalChecked(t *testing.T) {
 func TestWalletQuotaFromDecimalStrict(t *testing.T) {
 	quota, err := WalletQuotaFromDecimalStrict(decimal.NewFromInt(4_294_500_000))
 	require.NoError(t, err)
-	assert.Equal(t, 4_294_500_000, quota)
+	assert.Equal(t, int64(4_294_500_000), quota)
 
 	quota, err = WalletQuotaFromDecimalStrict(decimal.NewFromInt(MaxWalletQuota))
 	require.NoError(t, err)
